DiseaseForecasterApp-main/requirements.txt
streamlit
joblib
numpy
pandas
torch
torchvision
pillow
transformer
